consolidate zone status calculators
