+++ web/src/features/fba/components/infoPanel/FireZoneUnitInfo.tsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import { Box, ListItem, ListItemIcon, Typography } from '@mui/material'
-import { ADVISORY_ORANGE_FILL, ADVISORY_RED_FILL } from 'features/fba/components/map/featureStylers'
 import { FireShapeAreaDetail } from 'api/fbaAPI'
 import { useTheme } from '@mui/material/styles'
 import { TRANSPARENT_COLOUR } from 'app/theme'
+import { calculateStatusColour } from '@/features/fba/calculateZoneStatus'
 
 interface FireZoneUnitInfoProps {
   advisoryThreshold: number
@@ -13,31 +13,6 @@ interface FireZoneUnitInfoProps {
 
 const FireZoneUnitInfo = ({ advisoryThreshold, fireZoneUnitName, fireZoneUnitDetails }: FireZoneUnitInfoProps) => {
   const theme = useTheme()
-  const calculateStatus = (details: FireShapeAreaDetail[]) => {
-    // Default is transparent
-    let status = TRANSPARENT_COLOUR
-
-    if (details.length === 0) {
-      return status
-    }
-
-    const advisoryThresholdDetail = details.find(detail => detail.threshold == 1)
-    const warningThresholdDetail = details.find(detail => detail.threshold == 2)
-    const advisoryPercentage = advisoryThresholdDetail?.elevated_hfi_percentage ?? 0
-    const warningPercentage = warningThresholdDetail?.elevated_hfi_percentage ?? 0
-
-    if (advisoryPercentage + warningPercentage > advisoryThreshold) {
-      // advisory color orange
-      status = ADVISORY_ORANGE_FILL
-    }
-
-    if (warningPercentage > advisoryThreshold) {
-      // advisory color red
-      status = ADVISORY_RED_FILL
-    }
-
-    return status
-  }
 
   return (
     <ListItem data-testid="fire-zone-unit-info" sx={{ paddingBottom: theme.spacing(1), paddingTop: '0px' }}>
@@ -45,7 +20,7 @@ const FireZoneUnitInfo = ({ advisoryThreshold, fireZoneUnitName, fireZoneUnitDet
         <Box
           data-testid="fire-zone-unit-info-swatch"
           sx={{
-            backgroundColor: calculateStatus(fireZoneUnitDetails),
+            backgroundColor: calculateStatusColour(fireZoneUnitDetails, advisoryThreshold, TRANSPARENT_COLOUR),
             border: '1px solid #777',
             height: '0.7rem',
             width: '1rem'
